Started working on airport and parking space repositories
& tested saving airports and parking spaces into the DB in order to observe the connection between the two entities

+++ backend/src/main/java/com/clockworkcode/sibiuairportparkingmanager/repository/CarRepository.java
@@ -3,5 +3,8 @@ package com.clockworkcode.sibiuairportparkingmanager.repository;
 import com.clockworkcode.sibiuairportparkingmanager.model.Car;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarRepository extends JpaRepository<Car,Integer> {
+public interface CarRepository extends JpaRepository<Car,Long> {
+
+
+
 }
